Add todos to mark tests to be ported/investigated for porting to new framework. Move common functionality to main_test
